+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/ui/helper/PostHelper.java
@@ -38,12 +38,12 @@ import static java.util.concurrent.TimeUnit.SECONDS;
 
 public class PostHelper {
     public static CharSequence prependIcon(
-            @NonNull Context context, @NonNull CharSequence total, @NonNull Bitmap bitmap, int height
+            @NonNull Context context, @NonNull CharSequence total, @NonNull Bitmap bitmap, float height
     ) {
         int width = (int) (height / (bitmap.getHeight() / (float) bitmap.getWidth()));
         SpannableString string = new SpannableString("  " + total);
         ImageSpan imageSpan = new ImageSpan(context, bitmap);
-        imageSpan.getDrawable().setBounds(0, 0, width, height);
+        imageSpan.getDrawable().setBounds(0, 0, width, (int) height);
         string.setSpan(imageSpan, 0, 1, Spanned.SPAN_INCLUSIVE_EXCLUSIVE);
         return string;
     }
